+++ src/main/java/org/mapfish/print/map/readers/KaMapCacheMapReader.java
@@ -99,7 +99,7 @@ public class KaMapCacheMapReader extends TileableMapReader {
         if (unitEnum == null) {
             throw new RuntimeException("Unknown unit: '" + units + "'");
         }
-        final int scale = context.getConfig().getBestScale(Math.max(
+        final double scale = context.getConfig().getBestScale(Math.max(
             (maxGeoX - minGeoX) / (DistanceUnit.PT.convertTo(w, unitEnum)),
             (maxGeoY- minGeoY) / (DistanceUnit.PT.convertTo(h, unitEnum))));
